Separates the UnsafeMinPriorityQueue from PerThreadContainer.
